Reorganize includes and introduce MultiPlots API
