+++ .gitignore
@@ -6,6 +6,7 @@ subprojects/cpp_utils
 subprojects/QCustomPlot
 subprojects/packagecache/*
 subprojects/hedley-*
+subprojects/**/*
 *.egg-info
 *.srctrl*
 *.kdev4
@@ -13,3 +14,5 @@ subprojects/hedley-*
 *.pyc
 *.swp
 *.user
+*.user*
+.cache/*
