+++ include/SciQLopPlots/Items/QCPItemRichText.hpp
